strategies/protection.py
# Full file path: /moneyverse/strategies/protection.py

from centralized_logger import CentralizedLogger
from web3 import Web3
import os

# Initialize Web3 connection
web3 = Web3(Web3.HTTPProvider(os.getenv("WEB3_PROVIDER")))
logger = CentralizedLogger()

def activate_partner_protection(partner_entry, attack_types):
    """
    Activates protection strategies for a specified partner entry.
    """
    logger.log_info(f"Activating protection for partner entry: {partner_entry}")
    success = True
    try:
        for attack_type in attack_types:
            if attack_type == "counter_sandwich":
                result = perform_counter_sandwich(partner_entry)
            elif attack_type == "front_run":
                result = perform_front_run(partner_entry)
            elif attack_type == "back_run":
                result = perform_back_run(partner_entry)
            elif attack_type == "liquidity_drain":
                result = perform_liquidity_drain(partner_entry)
            else:
                logger.log_warning(f"Unknown protection strategy: {attack_type}")
                continue

            if result:
                logger.log_info(f"{attack_type.capitalize()} strategy executed successfully for {partner_entry}")
            else:
                logger.log_warning(f"{attack_type.capitalize()} strategy failed for {partner_entry}")
                success = False
    except Exception as e:
        logger.log_error(f"Failed to activate protection for {partner_entry}: {e}")
        success = False

    return success

# --- MEV Protection Strategies with Web3.py ---

def perform_counter_sandwich(entry):
    """
    Detects and foils sandwich attacks by placing counter transactions.
    """
    logger.log_info(f"Initiating counter-sandwich strategy on {entry}")
    sandwich_opportunity = detect_sandwich_attack(entry)
    
    if sandwich_opportunity:
        place_counter_sandwich(entry, sandwich_opportunity)
        logger.log_info(f"Counter-sandwich strategy completed for {entry}")
        return True
    else:
        logger.log_info(f"No sandwich attack detected on {entry}.")
        return False

def perform_front_run(entry):
    """
    Executes a front-running transaction to preemptively secure a transaction.
    """
    logger.log_info(f"Initiating front-run defense for {entry}")
    front_run_opportunity = detect_mempool_opportunity(entry)
    
    if front_run_opportunity:
        place_front_run(entry, front_run_opportunity)
        logger.log_info(f"Front-running defense executed for {entry}")
        return True
    else:
        logger.log_info(f"No front-run opportunities found on {entry}.")
        return False

def perform_back_run(entry):
    """
    Executes a back-running transaction to profit off a detected opportunity.
    """
    logger.log_info(f"Initiating back-run defense for {entry}")
    back_run_opportunity = detect_profitable_tx(entry)
    
    if back_run_opportunity:
        place_back_run(entry, back_run_opportunity)
        logger.log_info(f"Back-running defense executed for {entry}")
        return True
    else:
        logger.log_info(f"No back-run opportunities detected on {entry}.")
        return False

def perform_liquidity_drain(entry):
    """
    Executes a liquidity drain to counter large liquidity removal attempts.
    """
    logger.log_info(f"Initiating liquidity drain strategy for {entry}")
    drain_opportunity = detect_liquidity_removal(entry)
    
    if drain_opportunity:
        place_liquidity_drain(entry, drain_opportunity)
        logger.log_info(f"Liquidity drain executed for {entry}")
        return True
    else:
        logger.log_info(f"No liquidity removal detected for {entry}.")
        return False

# --- Real Logic Functions Using Web3.py ---

def detect_sandwich_attack(entry):
    """
    Detects sandwich attack attempts by analyzing the mempool and transaction patterns.
    """
    pending_txs = web3.eth.get_filter_logs("pending")  # Simulate mempool access for pending transactions
    for tx in pending_txs:
        if is_sandwich_pattern(tx, entry):
            return tx
    return None

def place_counter_sandwich(entry, tx_to_disrupt):
    """
    Places a transaction to disrupt an ongoing sandwich attack.
    """
    # Place a strategic transaction to disrupt the attacker’s pattern
    tx = {
        "to": entry,
        "value": Web3.toWei(0.1, "ether"),
        "gas": 300000,
        "gasPrice": tx_to_disrupt["gasPrice"] + Web3.toWei(1, "gwei")
    }
    signed_tx = web3.eth.account.sign_transaction(tx, private_key=os.getenv("PRIVATE_KEY"))
    web3.eth.send_raw_transaction(signed_tx.rawTransaction)
    logger.log_info(f"Counter-sandwich transaction sent to disrupt attack on {entry}")

def detect_mempool_opportunity(entry):
    """
    Detects front-running opportunities by scanning mempool for target transactions.
    """
    pending_txs = web3.eth.get_filter_logs("pending")
    for tx in pending_txs:
        if tx["to"] == entry:
            return tx
    return None

def place_front_run(entry, tx_to_preempt):
    """
    Executes a front-run transaction with higher gas price to ensure priority.
    """
    tx = {
        "to": entry,
        "value": Web3.toWei(0.1, "ether"),
        "gas": 300000,
        "gasPrice": tx_to_preempt["gasPrice"] + Web3.toWei(2, "gwei")
    }
    signed_tx = web3.eth.account.sign_transaction(tx, private_key=os.getenv("PRIVATE_KEY"))
    web3.eth.send_raw_transaction(signed_tx.rawTransaction)
    logger.log_info(f"Front-run transaction placed successfully for {entry}")

def detect_profitable_tx(entry):
    """
    Detects a profitable transaction worth back-running.
    """
    pending_txs = web3.eth.get_filter_logs("pending")
    for tx in pending_txs:
        if tx["to"] == entry and is_profitable(tx):
            return tx
    return None

def place_back_run(entry, tx_to_follow):
    """
    Places a back-run transaction to capitalize on the detected transaction.
    """
    tx = {
        "to": entry,
        "value": Web3.toWei(0.1, "ether"),
        "gas": 300000,
        "gasPrice": tx_to_follow["gasPrice"]
    }
    signed_tx = web3.eth.account.sign_transaction(tx, private_key=os.getenv("PRIVATE_KEY"))
    web3.eth.send_raw_transaction(signed_tx.rawTransaction)
    logger.log_info(f"Back-run transaction placed successfully for {entry}")

def detect_liquidity_removal(entry):
    """
    Detects liquidity removal attempts by monitoring liquidity pool contract events.
    """
    # Placeholder for detecting large liquidity withdrawals
    logs = web3.eth.get_logs({"fromBlock": "pending", "address": entry})
    for log in logs:
        if is_liquidity_removal(log):
            return log
    return None

def place_liquidity_drain(entry, log):
    """
    Places transactions in response to a detected liquidity drain attempt.
    """
    tx = {
        "to": entry,
        "value": Web3.toWei(0.2, "ether"),
        "gas": 500000,
        "gasPrice": Web3.toWei(20, "gwei")
    }
    signed_tx = web3.eth.account.sign_transaction(tx, private_key=os.getenv("PRIVATE_KEY"))
    web3.eth.send_raw_transaction(signed_tx.rawTransaction)
    logger.log_info(f"Liquidity drain transaction placed to protect {entry}")
